@@ -1,20 +1,18 @@
-import { SeoService } from './../../shared/seo.service';
-import { Component, OnInit } from '@angular/core';
+import { SeoService } from "../../shared/services/seo.service";
+import { Component, OnInit } from "@angular/core";
 
 @Component({
-  selector: 'app-data-protection',
-  templateUrl: './data-protection.component.html',
-  styleUrls: ['./data-protection.component.scss']
+  selector: "app-data-protection",
+  templateUrl: "./data-protection.component.html",
+  styleUrls: ["./data-protection.component.scss"],
 })
 export class DataProtectionComponent implements OnInit {
-
-  constructor(private seo: SeoService) { }
+  constructor(private seo: SeoService) {}
 
   ngOnInit() {
     this.seo.generateTags({
-      title: 'Datenschutz',
-      description: 'Datenschutzverordnung'
+      title: "Datenschutz",
+      description: "Datenschutzverordnung",
     });
   }
-
 }
