@@ -1,16 +1,12 @@
-import { Component, OnInit } from '@angular/core';
+import { Component, OnInit } from "@angular/core";
 
 @Component({
-  selector: 'app-outfitter-footer',
-  templateUrl: './outfitter-footer.component.html',
-  styleUrls: ['./outfitter-footer.component.scss']
+  selector: "app-outfitter-footer",
+  templateUrl: "./outfitter-footer.component.html",
+  styleUrls: ["./outfitter-footer.component.scss"],
 })
 export class OutfitterFooterComponent implements OnInit {
+  constructor() {}
 
-  constructor() { }
-
-  ngOnInit() {
-
-  }
-
+  ngOnInit() {}
 }
