@@ -1,15 +1,15 @@
-import { Component, OnInit } from '@angular/core';
-import { Post } from '@wh-objects/post';
-import { ActivatedRoute } from '@angular/router';
-import { TeamService } from '@wh-share/team.service';
-import { SeoService } from '@wh-share/seo.service';
-import { Paginator } from '@wh-objects/pagination';
-import { Team } from '@wh-objects/team';
+import { Component, OnInit } from "@angular/core";
+import { Post } from "@wh-share/objects/post";
+import { ActivatedRoute } from "@angular/router";
+import { TeamService } from "@wh-share/services/team.service";
+import { SeoService } from "@wh-share/services/seo.service";
+import { Paginator } from "@wh-share/objects/pagination";
+import { Team } from "@wh-share/objects/team";
 
 @Component({
-  selector: 'app-team-reports',
-  templateUrl: './team-reports.component.html',
-  styleUrls: ['./team-reports.component.scss']
+  selector: "app-team-reports",
+  templateUrl: "./team-reports.component.html",
+  styleUrls: ["./team-reports.component.scss"],
 })
 export class TeamReportsComponent implements OnInit {
   posts: Post[];
@@ -21,7 +21,11 @@ export class TeamReportsComponent implements OnInit {
   paginator: Paginator;
   myData: any;
 
-  constructor(private route: ActivatedRoute, private teamService: TeamService, private seo: SeoService) {
+  constructor(
+    private route: ActivatedRoute,
+    private teamService: TeamService,
+    private seo: SeoService
+  ) {
     // init data to hide console errors if nothing is found
     this.posts = new Array();
     this.team = new Team();
@@ -30,35 +34,45 @@ export class TeamReportsComponent implements OnInit {
   }
 
   private callApi() {
+    this.myData = this.route.params.subscribe(
+      (params) => {
+        this.teamID = params["id"];
 
-    this.myData = this.route.params.subscribe(params => {
-      this.teamID = params['id'];
-
-      this.teamService.getTeam(this.teamID).subscribe(
-        team => this.team = team,
-        error => { console.log(error); },
-        () => {
-          this.teamService.getTeamReports(this.team.wp.cat, this.page)
-            .subscribe(posts => this.posts = posts,
-              error => console.log('Error: ', error),
-              () => {
-                this.maxPages = this.teamService.getMaxPages();
-                this.pages = new Array();
-                for (let i = 1; i <= this.maxPages; i++) {
-                  if (i === this.page) {
-                    this.paginator = { id: i, active: 'active' };
-                  } else {
-                    this.paginator = { id: i, active: '' };
+        this.teamService.getTeam(this.teamID).subscribe(
+          (team) => (this.team = team),
+          (error) => {
+            console.log(error);
+          },
+          () => {
+            this.teamService
+              .getTeamReports(this.team.wp.cat, this.page)
+              .subscribe(
+                (posts) => (this.posts = posts),
+                (error) => console.log("Error: ", error),
+                () => {
+                  this.maxPages = this.teamService.getMaxPages();
+                  this.pages = new Array();
+                  for (let i = 1; i <= this.maxPages; i++) {
+                    if (i === this.page) {
+                      this.paginator = { id: i, active: "active" };
+                    } else {
+                      this.paginator = { id: i, active: "" };
+                    }
+                    this.pages.push(this.paginator);
                   }
-                  this.pages.push(this.paginator);
                 }
-              });
-        });
-    }, error => { console.log(error); }, () => { });
+              );
+          }
+        );
+      },
+      (error) => {
+        console.log(error);
+      },
+      () => {}
+    );
   }
 
-  ngOnInit() {
-  }
+  ngOnInit() {}
 
   callMethod(method: string) {
     this[method]();
@@ -67,6 +81,4 @@ export class TeamReportsComponent implements OnInit {
   setPages(value: number) {
     this.page = value;
   }
-
-
 }
