@@ -1,21 +1,20 @@
-import { Statistik } from './../../objects/hvw';
-import { Component, OnInit } from '@angular/core';
-import { HvwService } from '@wh-share/hvw.service';
-import { Ligue } from '@wh-objects/hvw';
-import { ActivatedRoute } from '@angular/router';
-import { SeoService } from '@wh-share/seo.service';
-import { TeamService } from '@wh-share/team.service';
-import { Team } from '@wh-objects/team';
-import { Globals } from '@wh-objects/globals';
-import { Observable } from 'rxjs';
+import { Statistik } from "../../shared/objects/hvw";
+import { Component, OnInit } from "@angular/core";
+import { HvwService } from "@wh-share/services/hvw.service";
+import { Ligue } from "@wh-share/objects/hvw";
+import { ActivatedRoute } from "@angular/router";
+import { SeoService } from "@wh-share/services/seo.service";
+import { TeamService } from "@wh-share/services/team.service";
+import { Team } from "@wh-share/objects/team";
+import { Globals } from "@wh-share/objects/globals";
+import { Observable } from "rxjs";
 
 @Component({
-  selector: 'app-team-result',
-  templateUrl: './team-result.component.html',
-  styleUrls: ['./team-result.component.scss']
+  selector: "app-team-result",
+  templateUrl: "./team-result.component.html",
+  styleUrls: ["./team-result.component.scss"],
 })
 export class TeamResultComponent implements OnInit {
-
   teamID: any;
   statTeam: string;
   stats: Statistik;
@@ -27,33 +26,44 @@ export class TeamResultComponent implements OnInit {
   buttonActive: Boolean = false;
   showSpinner = false;
 
-  constructor(private route: ActivatedRoute, private hvw: HvwService,
-    teams: TeamService, private seo: SeoService, private global: Globals) {
+  constructor(
+    private route: ActivatedRoute,
+    private hvw: HvwService,
+    teams: TeamService,
+    private seo: SeoService,
+    private global: Globals
+  ) {
     this.team = new Team();
-    this.teamID = '';
+    this.teamID = "";
 
-
-    this.sub = this.route.params.subscribe(params => {
-      this.teamID = params['id'];
+    this.sub = this.route.params.subscribe((params) => {
+      this.teamID = params["id"];
       teams.getTeam(this.teamID).subscribe(
-        team => this.team = team,
-        error => { console.log(error); },
+        (team) => (this.team = team),
+        (error) => {
+          console.log(error);
+        },
         () => {
           this.hvw.liga = this.team.hvw.liga;
           this.getApiData();
           this.changeLigueParams();
-        });
+        }
+      );
     });
   }
 
   private getApiData() {
     this.myHVW = this.hvw.getLigueData().subscribe(
-      ligue => this.ligue = ligue,
-      error => { console.log(error); },
+      (ligue) => (this.ligue = ligue),
+      (error) => {
+        console.log(error);
+      },
       () => {
         this.showSpinner = false;
         if (this.ligue.games !== undefined) {
-          this.ligue.games = this.ligue.games.filter((x) => this.isClub(x.team.home) || this.isClub(x.team.guest));
+          this.ligue.games = this.ligue.games.filter(
+            (x) => this.isClub(x.team.home) || this.isClub(x.team.guest)
+          );
         }
         if (this.ligue.scores !== undefined) {
           const myTeam = this.ligue.scores.filter((f) => {
@@ -68,7 +78,8 @@ export class TeamResultComponent implements OnInit {
           title: this.ligue.name,
           description: this.ligue.headline2,
         });
-      });
+      }
+    );
   }
 
   private changeLigueParams() {
@@ -94,11 +105,11 @@ export class TeamResultComponent implements OnInit {
     }
 
     this.hvw.liga = id;
-    this.secondLigueText = text + ' wechseln';
+    this.secondLigueText = text + " wechseln";
   }
 
   changeTeam(id: string) {
-    const selectedTeam = this.ligue.scores.find(element => element.id === id);
+    const selectedTeam = this.ligue.scores.find((element) => element.id === id);
     this.stats = selectedTeam.statistics;
     this.statTeam = selectedTeam.name;
   }
@@ -116,15 +127,15 @@ export class TeamResultComponent implements OnInit {
     let init;
     if (this.team.hvw.quali != null) {
       id = this.team.hvw.quali;
-      init = 'Zur Qualifikation';
+      init = "Zur Qualifikation";
     } else if (this.team.hvw.pokal != null) {
       id = this.team.hvw.pokal;
-      init = 'Zum Pokal';
+      init = "Zum Pokal";
     } else {
-      init = '';
+      init = "";
       this.buttonActive = true;
     }
-    return { id: id, text: 'Zur Saison', init: init };
+    return { id: id, text: "Zur Saison", init: init };
   }
 
   changeLigue() {
@@ -133,8 +144,7 @@ export class TeamResultComponent implements OnInit {
     this.getApiData();
   }
 
-  ngOnInit() {
-  }
+  ngOnInit() {}
 
   isClub(value): Boolean {
     return this.global.isOwnClub(value);
